requirements_py3.txt
asn1crypto==0.23.0
beautifulsoup4==4.6.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.1
future==0.16.0
idna==2.6
pony==0.7.2
psycopg2==2.7.3.1
pycparser==2.18
pyOpenSSL==17.3.0
python-telegram-bot==8.1.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
